added column block descriptor api and test fixes
