+++ crates/grafeo-core/src/codec/mod.rs
@@ -31,6 +31,7 @@
 
 pub mod bitpack;
 pub mod bitvec;
+pub mod block;
 pub mod delta;
 pub mod dictionary;
 #[cfg(feature = "tiered-storage")]
@@ -43,6 +44,7 @@ pub mod succinct;
 // Re-export commonly used types
 pub use bitpack::{BitPackedInts, DeltaBitPacked};
 pub use bitvec::BitVector;
+pub use block::BlockEntry;
 pub use delta::{DeltaEncoding, zigzag_decode, zigzag_encode};
 pub use dictionary::{DictionaryBuilder, DictionaryEncoding};
 pub use runlength::{Run, RunLengthAnalyzer, RunLengthEncoding, SignedRunLengthEncoding};
